Recipe.md
Recipe for Idiyappam ;
It is made of rice flour or ragi flour (Finger millet flour), salt and water. It is generally served as the main course at breakfast or dinner together with a curry (potato, egg, fish or meat curry) and coconut chutney. It is served with coconut milk and sugar in the Malabar region of Kerala. It is not usually served at lunch. In other parts of Kerala, Sri Lanka and Tamil Nadu, it is mostly eaten with spicy curries. Using wheat flour in its preparation gives it a brownish hue.

Mix rice flour with hot water, optionally add ghee, season with salt. Knead into a smooth dough. Fill an 'idiyappam' press or a sieve with the dough and press the noodles onto banana leaves or directly into an idli steamer. Add a little grated coconut if desired. Steam for 5–10 minutes. The idiyappam is served with coconut gratings and coconut milk.

The process for making putu mayam consists of mixing rice flour or idiyappam flour with water and/or coconut milk, and pressing the dough through a sieve to make vermicelli-like noodles. These are steamed, usually with the addition of juice from the aromatic pandan leaf (screwpine) as flavouring. The noodles are served with grated coconut and jaggery, or, preferably, gur (date palm sugar). In some areas, gula melaka (coconut palm sugar) is the favourite sweetener. Putu piring is a Malaysian version of putu mayam in which the rice flour dough is used to form a small cake around a filling of coconut and gur or jaggery.

Recipe of coconut chutney;
Now for the chutney take the coconut, baby onions, green chillies and red chilli powder & salt and grind it to a smooth paste. Then transfer it into a bowl. Add sufficient water and bring to required consistency.Then take a tadka pan and put oil. When hot add oil and then put the mustard seeds, dry red chilli and curry leaves. And after it splutters pour this over the chutney. 

After you do this your dish is ready to serve ;
You should mix it well with Idiyappam with oil so that it will not became like rice balls;
Try it ;
Thank you ;